avoid segfault in `eraser_unlink_kprobe_entry()`
The API for `check_sticky()` changed sometime between Linux kernel 4.7
and 6.13 and added an additional `mnt_idmap` argument. It isn't clear
where the `mnt_idmap` should come from, but it makes most sense for it
to belong to `regs->di`. We just need to be careful when traversing
filesystem structures to get to the `mnt_idmap` so as not to derefence a
null pointer.